+++ solver_rust_conic/totsu_f32cuda/CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 ### Changed
-- Update version of `libc`.
+- Update versions of `libc` and `totsu_core`.
 - Update versions of `totsu`, `float_eq` and `env_logger`.
 ### Deprecated
 ### Removed