+++ solver_rust_conic/totsu_f32cuda/CHANGELOG.md
@@ -16,10 +16,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 ### Security
 
+## [0.1.1] - 2023-01-03
+### Changed
+- Update versions of `libc` and `totsu_core`.
+- Update versions of `totsu`, `float_eq` and `env_logger`.
+
 ## [0.1.0] - 2022-09-23
 ### Changed
 - Divided into 4 crates.
 
 
-[unreleased]: https://github.com/convexbrain/Totsu/compare/totsu_f32cuda_v0.1.0...HEAD
+[unreleased]: https://github.com/convexbrain/Totsu/compare/totsu_f32cuda_v0.1.1...HEAD
+[0.1.1]: https://github.com/convexbrain/Totsu/releases/tag/totsu_f32cuda_v0.1.1
 [0.1.0]: https://github.com/convexbrain/Totsu/releases/tag/totsu_f32cuda_v0.1.0